amendments test split into local and remote api wrappers
